requirements.txt
numpy
pandas
scikit-learn
networkx
torch
transformers
spacy
pytest
pyyaml
requests
tokencost